pyproject.toml
[tool.poetry]
name = "pia"
version = "1.0.0"
description = "Agents for all purposes."
authors = ["alexander-fischer <alexfi@pm.me>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
typer = {extras = ["all"], version = "^0.12.3"}
setuptools = "^71.0.3"
python-hue-v2 = "^2.0.1"
python-dotenv = "^1.0.1"
speechrecognition = "^3.10.4"
loguru = "^0.7.2"
faster-whisper = "^1.0.3"
sounddevice = "^0.5.1"
kokoro-onnx = "^0.4.5"
nemo-toolkit = {extras = ["asr"], version = "^2.2.0"}
python-i18n = {extras = ["yaml"], version = "^0.3.9"}
asyncio = "^3.4.3"
pyaudio = "^0.2.14"
librosa = "^0.11.0"
onnx-asr = {extras = ["cpu", "hub"], version = "^0.6.0"}
lhotse = "1.30.0"
openwakeword = "^0.6.0"
openai = "^1.96.1"
piper-tts = "^1.3.0"

[tool.poetry.scripts]
pia = "pia.main:start_cli"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
